Renamed HCD module on store to hcdWallet
